Refactored code.| #2050
